refactor(core): scaffold honeybeepf-llm-core crate and move pure leaf modules
Introduce honeybeepf-llm-core, a pure-logic crate without aya, tokio,
OpenTelemetry, or /proc dependencies. It can be built and tested on any
platform without an eBPF toolchain, which speeds up developer iteration
and unblocks cross-platform CI for the analysis primitives.

This first PR moves the leaf modules:

- honeybeepf-llm/src/probes/builtin/llm/types.rs
  -> honeybeepf-llm-core/src/types.rs
- honeybeepf-llm/src/probes/builtin/llm/http/utils.rs
  -> honeybeepf-llm-core/src/byte_utils.rs
- honeybeepf-llm/src/probes/builtin/llm/http/providers/usage.rs::get_nested_value
  -> honeybeepf-llm-core/src/byte_utils.rs (lifted to pub, shared utility)

Tests:
- get_nested_value test moves to honeybeepf-llm-core/tests/byte_utils.rs
- All other inline tests in the agent crate remain in place; subsequent PRs
  will move provider/protocol/processor modules with their own tests.

CI:
- New core-test job (Ubuntu + macOS) runs cargo test -p honeybeepf-llm-core
  using stable Rust, no eBPF toolchain required.

Refs #12
