+++ .github/workflows/docker-publish.yml
@@ -34,6 +34,31 @@ jobs:
         working-directory: honeybeepf-llm
         run: cargo fmt --all -- --check
 
+  core-test:
+    name: Core lib (${{ matrix.os }})
+    needs: lint
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Install Rust toolchain (stable)
+        uses: dtolnay/rust-toolchain@stable
+
+      - name: Cache cargo
+        uses: Swatinem/rust-cache@v2
+        with:
+          workspaces: honeybeepf-llm
+          cache-on-failure: true
+
+      - name: Test core crate
+        working-directory: honeybeepf-llm
+        run: cargo test -p honeybeepf-llm-core
+
   build-and-test:
     name: Build (Rust) and sanity-check
     needs: lint
