+++ honeybeepf-llm/honeybeepf-llm-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "honeybeepf-llm-core"
+version = "0.1.0"
+edition.workspace = true
+
+license.workspace = true
+
+[dependencies]
+honeybeepf-llm-common = { path = "../honeybeepf-llm-common" }
+
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+
+[lib]
+path = "src/lib.rs"
