+++ honeybeepf-llm/honeybeepf-llm/src/probes/builtin/llm/processor.rs
@@ -1,11 +1,9 @@
 use std::time::Instant;
 
+use honeybeepf_llm_core::types::LlmDirection;
 use log::{info, warn};
 
-use crate::probes::builtin::llm::{
-    http::{self, ProtocolParser},
-    types::LlmDirection,
-};
+use crate::probes::builtin::llm::http::{self, ProtocolParser};
 
 // Buffer size constants
 const INITIAL_BUFFER_CAPACITY: usize = 8 * 1024; // 8KB initial allocation
